Photos showing the Hesse with Choc v2 switches and Tai-Hao THCS keycaps.
Tai-Hao sell these very flat keycaps with either the cross stem
(used here with choc v2 switches) or the two legs (choc v1). They
look very smart, although not quite the color I expected - to me
these 'Navy' ones are a green/gray not a blue/gray as in their photos.
Because the surface is flat and unsculpted, it is easy to get your
fingers out of position - but at least the Gamma Omega is a small
board. Sadly they do not (yet) sell matching homing keys, but on
the bright side two packs of 18 is just right for a 36-key keyboard.

The switches are 35g Kailh Islet Pink Mini Low Profile Silent Linear
Switches, and they are very stable and very quiet. Quite a change
from the comparatively noisy 50g Kailh Choc 'red' switches I had in
before, which are noisier and rattle, and which now feel a little
too heavy.